+++ forui/lib/forui.dart
@@ -31,6 +31,7 @@ export 'widgets/progress.dart';
 export 'widgets/radio.dart';
 export 'widgets/resizable.dart';
 export 'widgets/scaffold.dart';
+export 'widgets/select.dart';
 export 'widgets/select_group.dart';
 export 'widgets/select_menu_tile.dart';
 export 'widgets/select_tile_group.dart';