+++ forui/lib/widgets/select.dart
@@ -1,0 +1,4 @@
+///
+library forui.widgets.select;
+
+export '../src/widgets/select/select_controller.dart';