+++ forui_hooks/test/src/select_group_controller_hook_test.dart
@@ -7,39 +7,14 @@ import 'package:forui/forui.dart';
 import 'package:forui_hooks/forui_hooks.dart';
 
 void main() {
-  testWidgets('useFRadioSelectGroupController', (tester) async {
-    late FRadioSelectGroupController<int> controller;
+  testWidgets('useFSelectController', (tester) async {
+    late FSelectController<int> controller;
 
     await tester.pumpWidget(
       MaterialApp(
         home: HookBuilder(
           builder: (context) {
-            controller = useFRadioSelectGroupController();
-            return FSelectGroup(
-              controller: controller,
-              items: const [
-                FSelectGroupItem.radio(value: 0, label: Text('0')),
-                FSelectGroupItem.radio(value: 1, label: Text('1')),
-              ],
-            );
-          },
-        ),
-      ),
-    );
-
-    await tester.tap(find.byType(FRadio).first);
-
-    await tester.pumpAndSettle();
-  });
-
-  testWidgets('useFMultiSelectGroupController', (tester) async {
-    late FMultiSelectGroupController<int> controller;
-
-    await tester.pumpWidget(
-      MaterialApp(
-        home: HookBuilder(
-          builder: (context) {
-            controller = useFMultiSelectGroupController();
+            controller = useFSelectController();
             return FSelectGroup(
               controller: controller,
               items: const [
@@ -53,6 +28,31 @@ void main() {
     );
 
     await tester.tap(find.byType(FCheckbox).first);
+
+    await tester.pumpAndSettle();
+  });
+
+  testWidgets('useFRadioSelectcontroller', (tester) async {
+    late FSelectController<int> controller;
+
+    await tester.pumpWidget(
+      MaterialApp(
+        home: HookBuilder(
+          builder: (context) {
+            controller = useFRadioSelectController();
+            return FSelectGroup(
+              controller: controller,
+              items: const [
+                FSelectGroupItem.radio(value: 0, label: Text('0')),
+                FSelectGroupItem.radio(value: 1, label: Text('1')),
+              ],
+            );
+          },
+        ),
+      ),
+    );
+
+    await tester.tap(find.byType(FRadio).first);
 
     await tester.pumpAndSettle();
   });